refactor: refactor validation function
